+++ src/QueryCat.Cli/Commands/ServeCommand.cs
@@ -12,18 +12,24 @@ internal class ServeCommand : BaseCommand
     public ServeCommand() : base("serve", "Run simple HTTP server.")
     {
         var urlsOption = new Option<string>("--url", description: "Endpoint to serve on.");
+        var allowOriginOption = new Option<string>("--allow-origin", description: "Enables CORS for the specified origin.");
 
         this.AddOption(urlsOption);
-        this.SetHandler((queryOptions, urls) =>
+        this.AddOption(allowOriginOption);
+        this.SetHandler((queryOptions, urls, allowOrigin) =>
         {
             var executionThread = queryOptions.CreateExecutionThread(new ExecutionOptions
             {
                 PagingSize = ExecutionOptions.NoLimit,
-                AddRowNumberColumn = false,
+                AddRowNumberColumn = true,
                 DefaultRowsOutput = NullRowsOutput.Instance,
             });
             var webServer = new WebServer(executionThread, urls);
+            if (!string.IsNullOrEmpty(allowOrigin))
+            {
+                webServer.AllowOrigin = allowOrigin;
+            }
             webServer.Run();
-        }, new ApplicationOptionsBinder(LogLevelOption, PluginDirectoriesOption), urlsOption);
+        }, new ApplicationOptionsBinder(LogLevelOption, PluginDirectoriesOption), urlsOption, allowOriginOption);
     }
 }